# Conflicts:
#	modules/control-utility/src/main/java/org/apache/ignite/internal/commandline/Command.java
#	modules/control-utility/src/test/java/org/apache/ignite/util/GridCommandHandlerTest.java

+++ modules/platforms/cpp/core/src/impl/interop/interop_target.cpp
@@ -104,7 +104,7 @@ namespace ignite
 
                 if (outPtr)
                 {
-                    long long res = env.Get()->Context()->TargetInStreamOutLong(javaRef, opType, outPtr, &jniErr);
+                    int64_t res = env.Get()->Context()->TargetInStreamOutLong(javaRef, opType, outPtr, &jniErr);
 
                     IgniteError::SetError(jniErr.code, jniErr.errCls, jniErr.errMsg, err);
 
@@ -125,7 +125,7 @@ namespace ignite
 
                 if (outPtr)
                 {
-                    long long res = env.Get()->Context()->TargetInStreamOutLong(javaRef, opType, outPtr, &jniErr);
+                    int64_t res = env.Get()->Context()->TargetInStreamOutLong(javaRef, opType, outPtr, &jniErr);
 
                     IgniteError::SetError(jniErr.code, jniErr.errCls, jniErr.errMsg, err);
 
@@ -140,7 +140,7 @@ namespace ignite
             {
                 JniErrorInfo jniErr;
 
-                long long res = env.Get()->Context()->TargetInLongOutLong(javaRef, opType, 0, &jniErr);
+                int64_t res = env.Get()->Context()->TargetInLongOutLong(javaRef, opType, 0, &jniErr);
 
                 IgniteError::SetError(jniErr.code, jniErr.errCls, jniErr.errMsg, err);
 
@@ -304,7 +304,7 @@ namespace ignite
             {
                 JniErrorInfo jniErr;
 
-                long long res = env.Get()->Context()->TargetInLongOutLong(javaRef, opType, val, &jniErr);
+                int64_t res = env.Get()->Context()->TargetInLongOutLong(javaRef, opType, val, &jniErr);
 
                 IgniteError::SetError(jniErr.code, jniErr.errCls, jniErr.errMsg, err);
